#12 The test for SmartFactory.createSOBjectList is also failing (reaching script line limit) when used for creating 200 objects (the common bulk size).
I tried to reduce the executed script lines to be able to create at
least 201 objects (only tested with contact objects - this is
strongly coupled to the size of available fields for a given type).

The main and most valuable changes are related to the setFieldValue
method. More concrete the isIncludedField, isExcludedField and
isPersonAccountField methods that will be called from the setFieldValue
method. There are the expressions are now inlined in the return
statements (readability will get worse through this changes).

Also the createSObjectList method and some related methods was adjusted
to safe a view lines during iterating over the number of objects
(mainly create Schema.objectType only once for bulk object creation).